Move regparser/tree/interp into interpparser
